refacto tests
