+++ tests/lists/create-list.spec.ts
@@ -1,27 +1,23 @@
 import { test, expect } from "tests/fixtures/auth";
 import { signin } from "tests/helpers/signin";
-import { fillListName } from "tests/helpers/fillListName";
+import { createList } from "tests/helpers/createList";
 
 test("Authenticated users can create lists", async ({ page, testUser }) => {
     // connexion
     await signin(page, testUser.email, testUser.password);
 
-    // check heading and button
+    // check heading
     const heading = page.getByRole("heading", { level: 1, name: /Mes listes/i });
     await expect(heading).toBeVisible();
 
+    // check button
     const button = page.getByRole("button", { name: /Créer une liste/i });
     await expect(button).toBeVisible();
 
-    // Create a list
-    await page.getByRole("button", { name: /Créer une liste/i }).click();
-    await page.waitForURL("/add-list");
-    const titleInput = page.getByLabel("Titre*");
-    await expect(titleInput).toBeVisible();
-    await fillListName(page);
-    await page.getByRole("button", { name: /Créer une liste/i }).click();
+    // create a list
+    await createList(page);
 
-    //Expects the new list is visible
-    const listItem = page.getByText("Playwright List");
-    await expect(listItem).toBeVisible();
+    // check list title
+    const headingList = page.getByRole("heading", { name: /Playwright List/i });
+    await expect(headingList).toBeVisible();
 });
