+++ tests/lists/edit-list.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "tests/fixtures/auth";
 import { signin } from "tests/helpers/signin";
-import { fillListName } from "tests/helpers/fillListName";
+import { createList } from "tests/helpers/createList";
 import { editListName } from "tests/helpers/editListName";
 
 test("Authenticated users can update lists", async ({ page, testUser }) => {
@@ -8,14 +8,7 @@ test("Authenticated users can update lists", async ({ page, testUser }) => {
     await signin(page, testUser.email, testUser.password);
 
     //create a list
-    await page.getByRole("button", { name: /Créer une liste/i }).click();
-    await page.waitForURL("/add-list");
-    await fillListName(page);
-    await page.getByRole("button", { name: /Créer une liste/i }).click();
-
-    // check list title
-    const heading = page.getByRole("heading", { name: /Playwright List/i });
-    await expect(heading).toBeVisible();
+    await createList(page);
 
     // navigate to edit list
     await page.getByRole("link", { name: /Modifier la liste/i }).click();
